+++ nata-monkey/src/main/java/com/nata/action/BackAction.java
@@ -22,6 +22,6 @@ public class BackAction extends Action{
 
     @Override
     public String toString() {
-        return "BackAction";
+        return super.toString()+"BackAction";
     }
 }