+++ nata-monkey/src/main/java/com/nata/action/LongClickAction.java
@@ -27,8 +27,38 @@ public class LongClickAction extends Action{
 
     @Override
     public String toString() {
-        return "LongClickAction{" +
+        return super.toString()+"LongClickAction{" +
                 "element=" + element +
                 '}';
+    }
+
+    @Override
+    public int hashCode() {
+        int hash = 17;
+        hash = 31 * hash + element.hashCode();
+        return hash;
+    }
+
+    @Override
+    public boolean equals(Object otherObject) {
+        if (this == otherObject) {
+            return true;
+        }
+
+        if (otherObject == null) {
+            return false;
+        }
+
+        if (getClass() != otherObject.getClass()) {
+            return false;
+        }
+
+        LongClickAction other = (LongClickAction) otherObject;
+
+        if(!element.equals(other.element)){
+            return false;
+        }
+
+        return true;
     }
 }