+++ nata-monkey/src/main/java/com/nata/action/TextInputAction.java
@@ -12,24 +12,63 @@ import com.nata.element.Element;
 public class TextInputAction extends Action{
     private AdbDevice device = null;
     private Element element = null;
+    private String text = null;
     public TextInputAction(AdbDevice device,Element element){
         super.setName(ActionType.INPUT);
         this.device = device;
         this.element = element;
+        this.text = TextValueDictionary.getRandomValidValue();
     }
 
     @Override
     public void fire() {
         device.tap(element.getX(),element.getY());
-        device.sendText(TextValueDictionary.getRandomValidValue());
+        device.sendText(text);
+        device.hideSoftKeyBoard();
         count++;
     }
-
 
     @Override
     public String toString() {
         return "TextInputAction{" +
                 "element=" + element +
+                ", text='" + text + '\'' +
                 '}';
     }
+
+    @Override
+    public int hashCode() {
+        int hash = 17;
+        hash = 31 * hash + element.hashCode();
+        hash = 31 * hash + text.hashCode();
+        return hash;
+    }
+
+    @Override
+    public boolean equals(Object otherObject) {
+        if (this == otherObject) {
+            return true;
+        }
+
+        if (otherObject == null) {
+            return false;
+        }
+
+        if (getClass() != otherObject.getClass()) {
+            return false;
+        }
+
+        TextInputAction other = (TextInputAction) otherObject;
+
+        if(!element.equals(other.element)){
+            return false;
+        }
+
+        if(!text.equals(other.text)){
+            return false;
+        }
+
+        return true;
+    }
+
 }